remove RankingPredictionWriter
